fix compiler warning
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@1643 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/src/univcsvdialog.cpp
@@ -573,7 +573,7 @@ void mmUnivCSVImportDialog::OnRemove(wxCommandEvent& /*event*/)
         csvListBox_->Delete(selIndex);
         csvFieldOrder_.erase(csvFieldOrder_.begin() + selIndex);
 
-        if (selIndex < csvListBox_->GetCount())
+        if (selIndex < (int)csvListBox_->GetCount())
             csvListBox_->SetSelection(selIndex, true);
         else
             csvListBox_->SetSelection(csvListBox_->GetCount() - 1, true);
